requirements.txt
arabic-reshaper==3.0.0
asgiref==3.4.1
asn1crypto==1.5.1
boto3==1.18.30
botocore==1.21.30
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-tailwind==0.5.0
cryptography==40.0.2
cssselect2==0.7.0
distlib==0.3.6
dj-database-url==0.5.0
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
django-widget-tweaks==1.4.8
filelock==3.12.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
jmespath==0.10.0
lxml==4.9.2
mysql-connector-python==8.0.31
oscrypto==1.3.0
Pillow==9.5.0
platformdirs==3.5.0
psycopg2==2.9.1
pycparser==2.21
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
pypdf==3.8.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.13
requests==2.30.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.1
svglib==1.5.1
tinycss2==1.2.1
typing-extensions==3.10.0.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==1.26.6
virtualenv==20.23.0
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.11
